toolchaingenericds branch that has stable gdb at arm7

+++ src/arm9/source/arm9_driver/videoTGDS.h
@@ -28,8 +28,11 @@ USA
 
 
 
-#define VRAM_ENABLE		(uint8)(1<<7)
-#define VRAM_OFFSET(n)	(uint8)((n)<<3)
+#define VRAM_ENABLE		(1<<7)
+#define VRAM_OFFSET(n)	((n)<<3)
+
+
+
 
 #define VRAM_MAX_LETTER_INDEX	(sint32)(9)
 
@@ -70,113 +73,113 @@ typedef struct vramSetup
   I       16K   0    -     68A0000h-68A3FFFh
  */
 
-#define VRAM_A_LCDC_MODE 	(uint8)(0)
-#define VRAM_B_LCDC_MODE 	(uint8)(0)
-#define VRAM_C_LCDC_MODE 	(uint8)(0)
-#define VRAM_D_LCDC_MODE 	(uint8)(0)
-#define VRAM_E_LCDC_MODE 	(uint8)(0)
-#define VRAM_F_LCDC_MODE 	(uint8)(0)
-#define VRAM_G_LCDC_MODE 	(uint8)(0)
-#define VRAM_H_LCDC_MODE 	(uint8)(0)
-#define VRAM_I_LCDC_MODE 	(uint8)(0)
+#define VRAM_A_LCDC_MODE 	(0)
+#define VRAM_B_LCDC_MODE 	(0)
+#define VRAM_C_LCDC_MODE 	(0)
+#define VRAM_D_LCDC_MODE 	(0)
+#define VRAM_E_LCDC_MODE 	(0)
+#define VRAM_F_LCDC_MODE 	(0)
+#define VRAM_G_LCDC_MODE 	(0)
+#define VRAM_H_LCDC_MODE 	(0)
+#define VRAM_I_LCDC_MODE 	(0)
 
 //VRAM Setup 1: //MST 1
 //ARM9, 2D Graphics Engine A, BG-VRAM (max 512K)
 //Vram A (128K)
-#define VRAM_A_0x06000000_ENGINE_A_BG 	(uint8)(1 | VRAM_OFFSET(0))	//OFS 0..3
-#define VRAM_A_0x06020000_ENGINE_A_BG 	(uint8)(1 | VRAM_OFFSET(1))
-#define VRAM_A_0x06040000_ENGINE_A_BG 	(uint8)(1 | VRAM_OFFSET(2))
-#define VRAM_A_0x06060000_ENGINE_A_BG 	(uint8)(1 | VRAM_OFFSET(3))
+#define VRAM_A_0x06000000_ENGINE_A_BG 	(1 | VRAM_OFFSET(0))	//OFS 0..3
+#define VRAM_A_0x06020000_ENGINE_A_BG 	(1 | VRAM_OFFSET(1))
+#define VRAM_A_0x06040000_ENGINE_A_BG 	(1 | VRAM_OFFSET(2))
+#define VRAM_A_0x06060000_ENGINE_A_BG 	(1 | VRAM_OFFSET(3))
 
 //Vram B (128K)
-#define VRAM_B_0x06000000_ENGINE_A_BG 	(uint8)(1 | VRAM_OFFSET(0))	//OFS 0..3
-#define VRAM_B_0x06020000_ENGINE_A_BG 	(uint8)(1 | VRAM_OFFSET(1))
-#define VRAM_B_0x06040000_ENGINE_A_BG 	(uint8)(1 | VRAM_OFFSET(2))
-#define VRAM_B_0x06060000_ENGINE_A_BG 	(uint8)(1 | VRAM_OFFSET(3))
+#define VRAM_B_0x06000000_ENGINE_A_BG 	(1 | VRAM_OFFSET(0))	//OFS 0..3
+#define VRAM_B_0x06020000_ENGINE_A_BG 	(1 | VRAM_OFFSET(1))
+#define VRAM_B_0x06040000_ENGINE_A_BG 	(1 | VRAM_OFFSET(2))
+#define VRAM_B_0x06060000_ENGINE_A_BG 	(1 | VRAM_OFFSET(3))
 
 //Vram C (128K)
-#define VRAM_C_0x06000000_ENGINE_A_BG 	(uint8)(1 | VRAM_OFFSET(0))	//OFS 0..3
-#define VRAM_C_0x06020000_ENGINE_A_BG 	(uint8)(1 | VRAM_OFFSET(1))
-#define VRAM_C_0x06040000_ENGINE_A_BG 	(uint8)(1 | VRAM_OFFSET(2))
-#define VRAM_C_0x06060000_ENGINE_A_BG 	(uint8)(1 | VRAM_OFFSET(3))
+#define VRAM_C_0x06000000_ENGINE_A_BG 	(1 | VRAM_OFFSET(0))	//OFS 0..3
+#define VRAM_C_0x06020000_ENGINE_A_BG 	(1 | VRAM_OFFSET(1))
+#define VRAM_C_0x06040000_ENGINE_A_BG 	(1 | VRAM_OFFSET(2))
+#define VRAM_C_0x06060000_ENGINE_A_BG 	(1 | VRAM_OFFSET(3))
 
 //Vram D (128K)
-#define VRAM_D_0x06000000_ENGINE_A_BG 	(uint8)(1 | VRAM_OFFSET(0))	//OFS 0..3
-#define VRAM_D_0x06020000_ENGINE_A_BG 	(uint8)(1 | VRAM_OFFSET(1))
-#define VRAM_D_0x06040000_ENGINE_A_BG 	(uint8)(1 | VRAM_OFFSET(2))
-#define VRAM_D_0x06060000_ENGINE_A_BG 	(uint8)(1 | VRAM_OFFSET(3))
+#define VRAM_D_0x06000000_ENGINE_A_BG 	(1 | VRAM_OFFSET(0))	//OFS 0..3
+#define VRAM_D_0x06020000_ENGINE_A_BG 	(1 | VRAM_OFFSET(1))
+#define VRAM_D_0x06040000_ENGINE_A_BG 	(1 | VRAM_OFFSET(2))
+#define VRAM_D_0x06060000_ENGINE_A_BG 	(1 | VRAM_OFFSET(3))
 
 //Vram E  64K   1    -     6000000h
-#define VRAM_E_0x06000000_ENGINE_A_BG 	(uint8)(1)		//static @ 0x06000000
+#define VRAM_E_0x06000000_ENGINE_A_BG 	(1)		//static @ 0x06000000
 
 //Vram F (16K), takes in account OFS0 and OFS1 : 06000000h+(4000h*OFS.0)+(10000h*OFS.1)
-#define VRAM_F_0x06000000_ENGINE_A_BG  	(uint8)(1 | (( 0 )<<3))	//OFS 0
-#define VRAM_F_0x06004000_ENGINE_A_BG 	(uint8)(1 | (( 1 )<<3))	//OFS 1
-#define VRAM_F_0x06010000_ENGINE_A_BG  	(uint8)(1 | (( 2 )<<3))	//OFS 2
-#define VRAM_F_0x06014000_ENGINE_A_BG  	(uint8)(1 | (( 3 )<<3))	//OFS 3
+#define VRAM_F_0x06000000_ENGINE_A_BG  	(1 | (( 0 )<<3))	//OFS 0
+#define VRAM_F_0x06004000_ENGINE_A_BG 	(1 | (( 1 )<<3))	//OFS 1
+#define VRAM_F_0x06010000_ENGINE_A_BG  	(1 | (( 2 )<<3))	//OFS 2
+#define VRAM_F_0x06014000_ENGINE_A_BG  	(1 | (( 3 )<<3))	//OFS 3
 
 //Vram G (16K),	takes in account OFS0 and OFS1 : 06000000h+(4000h*OFS.0)+(10000h*OFS.1)
-#define VRAM_G_0x060XXXXX_ENGINE_A_BG 	(uint8)(1)	//OFS 0..3
+#define VRAM_G_0x060XXXXX_ENGINE_A_BG 	(1)	//OFS 0..3
 
 
 //VRAM Setup 2: //MST 2
 //ARM9, 2D Graphics Engine A, OBJ-VRAM (max 256K) (OFS 0..1)
-#define VRAM_A_0x06400000_ENGINE_A_BG 	(uint8)(2 | VRAM_OFFSET(0))	//0x064(20000*OFS.0) (OFS.1 always zero)
-#define VRAM_A_0x06420000_ENGINE_A_BG 	(uint8)(2 | VRAM_OFFSET(1))	
+#define VRAM_A_0x06400000_ENGINE_A_BG 	(2 | VRAM_OFFSET(0))	//0x064(20000*OFS.0) (OFS.1 always zero)
+#define VRAM_A_0x06420000_ENGINE_A_BG 	(2 | VRAM_OFFSET(1))	
 
-#define VRAM_B_0x06400000_ENGINE_A_BG 	(uint8)(2 | VRAM_OFFSET(0))	//0x064(20000*OFS.0) (OFS.1 always zero)
-#define VRAM_B_0x06420000_ENGINE_A_BG 	(uint8)(2 | VRAM_OFFSET(1))	
+#define VRAM_B_0x06400000_ENGINE_A_BG 	(2 | VRAM_OFFSET(0))	//0x064(20000*OFS.0) (OFS.1 always zero)
+#define VRAM_B_0x06420000_ENGINE_A_BG 	(2 | VRAM_OFFSET(1))	
 
 //both can coexist through OFS 
-#define VRAM_E_0x06400000_ENGINE_A_BG 	(uint8)(2)	//static	: VRAM_E: Main Engine 0x06400000 64K MST 2
+#define VRAM_E_0x06400000_ENGINE_A_BG 	(2)	//static	: VRAM_E: Main Engine 0x06400000 64K MST 2
 
 //Vram F (16K), takes in account OFS0 and OFS1 : 6400000h+(4000h*OFS.0)+(10000h*OFS.1)
-#define VRAM_F_0x06400000_ENGINE_A_BG  (uint8)(2 | (( 0 )<<3))	//OFS 0
-#define VRAM_F_0x06404000_ENGINE_A_BG  (uint8)(2 | (( 1 )<<3))	//OFS 1
-#define VRAM_F_0x06410000_ENGINE_A_BG  (uint8)(2 | (( 2 )<<3))	//OFS 2
-#define VRAM_F_0x06414000_ENGINE_A_BG  (uint8)(2 | (( 3 )<<3))	//OFS 3
+#define VRAM_F_0x06400000_ENGINE_A_BG  (2 | (( 0 )<<3))	//OFS 0
+#define VRAM_F_0x06404000_ENGINE_A_BG  (2 | (( 1 )<<3))	//OFS 1
+#define VRAM_F_0x06410000_ENGINE_A_BG  (2 | (( 2 )<<3))	//OFS 2
+#define VRAM_F_0x06414000_ENGINE_A_BG  (2 | (( 3 )<<3))	//OFS 3
 
 //Vram G (16K),	takes in account OFS0 and OFS1 : 6400000h+(4000h*OFS.0)+(10000h*OFS.1)
-#define VRAM_G_0x06400000_ENGINE_A_BG	(uint8)(2 | (( 0 )<<3))
-#define VRAM_G_0x06404000_ENGINE_A_BG	(uint8)(2 | (( 1 )<<3))
-#define VRAM_G_0x06410000_ENGINE_A_BG	(uint8)(2 | (( 2 )<<3))
-#define VRAM_G_0x06414000_ENGINE_A_BG	(uint8)(2 | (( 3 )<<3))
+#define VRAM_G_0x06400000_ENGINE_A_BG	(2 | (( 0 )<<3))
+#define VRAM_G_0x06404000_ENGINE_A_BG	(2 | (( 1 )<<3))
+#define VRAM_G_0x06410000_ENGINE_A_BG	(2 | (( 2 )<<3))
+#define VRAM_G_0x06414000_ENGINE_A_BG	(2 | (( 3 )<<3))
 
 //VRAM Setup 3
 //2D Graphics Engine A, BG Extended Palette. This mode maps memory to VRAM and not CPU addresses
 //Vram E (64K) : Slot 0-3  ;only lower 32K used (No OFS)
-#define VRAM_E_SLOT0_ENGINE_A_BG_EXTENDED 	(uint8)(4 | (( 0 )<<3))	
-#define VRAM_E_SLOT1_ENGINE_A_BG_EXTENDED 	(uint8)(4 | (( 1 )<<3))	
-#define VRAM_E_SLOT2_ENGINE_A_BG_EXTENDED 	(uint8)(4 | (( 2 )<<3))	
-#define VRAM_E_SLOT3_ENGINE_A_BG_EXTENDED 	(uint8)(4 | (( 3 )<<3))
+#define VRAM_E_SLOT0_ENGINE_A_BG_EXTENDED 	(4 | (( 0 )<<3))	
+#define VRAM_E_SLOT1_ENGINE_A_BG_EXTENDED 	(4 | (( 1 )<<3))	
+#define VRAM_E_SLOT2_ENGINE_A_BG_EXTENDED 	(4 | (( 2 )<<3))	
+#define VRAM_E_SLOT3_ENGINE_A_BG_EXTENDED 	(4 | (( 3 )<<3))
 
 //Vram F (16K): ofst 0..1
 //Vram G (16K): ofst 0..1
-#define VRAM_F_SLOT01_ENGINE_A_BG_EXTENDED 	(uint8)(4 | (( 0 )<<3))	  	//Slot 0-1 (OFS=0)
-#define VRAM_F_SLOT23_ENGINE_A_BG_EXTENDED 	(uint8)(4 | (( 1 )<<3))		//Slot 2-3 (OFS=1)
-#define VRAM_G_SLOT01_ENGINE_A_BG_EXTENDED 	(uint8)(4 | (( 0 )<<3))	  	//Slot 0-1 (OFS=0)
-#define VRAM_G_SLOT23_ENGINE_A_BG_EXTENDED 	(uint8)(4 | (( 1 )<<3))		//Slot 2-3 (OFS=1)
+#define VRAM_F_SLOT01_ENGINE_A_BG_EXTENDED 	(4 | (( 0 )<<3))	  	//Slot 0-1 (OFS=0)
+#define VRAM_F_SLOT23_ENGINE_A_BG_EXTENDED 	(4 | (( 1 )<<3))		//Slot 2-3 (OFS=1)
+#define VRAM_G_SLOT01_ENGINE_A_BG_EXTENDED 	(4 | (( 0 )<<3))	  	//Slot 0-1 (OFS=0)
+#define VRAM_G_SLOT23_ENGINE_A_BG_EXTENDED 	(4 | (( 1 )<<3))		//Slot 2-3 (OFS=1)
 
 
 //VRAM Setup 7
 //Vram C (128K)
-#define VRAM_C_0x06200000_ENGINE_B_BG 	(uint8)(4 | VRAM_OFFSET(0))	//static
+#define VRAM_C_0x06200000_ENGINE_B_BG 	(4 | VRAM_OFFSET(0))	//static
 
 //Vram H (32K)
-#define VRAM_H_0x06200000_ENGINE_B_BG 	(uint8)(1)	//6200000
+#define VRAM_H_0x06200000_ENGINE_B_BG 	(1)	//6200000
 
 //Vram I (16K)
-#define VRAM_I_0x06208000_ENGINE_B_BG 	(uint8)(1)	//static. 
+#define VRAM_I_0x06208000_ENGINE_B_BG 	(1)	//static. 
 
 
 //VRAM Setup 11
 
 //Vram C,D     128K  2    0..1  6000000h+(20000h*OFS.0)  ;OFS.1 must be zero
-#define VRAM_C_0x06000000_ARM7 	(uint8)(2 | VRAM_OFFSET(0))	//static
-#define VRAM_C_0x06020000_ARM7 	(uint8)(2 | VRAM_OFFSET(1))	//static
+#define VRAM_C_0x06000000_ARM7 	(2 | VRAM_OFFSET(0))	//static
+#define VRAM_C_0x06020000_ARM7 	(2 | VRAM_OFFSET(1))	//static
 
-#define VRAM_D_0x06000000_ARM7 	(uint8)(2 | VRAM_OFFSET(0))	//static
-#define VRAM_D_0x06020000_ARM7 	(uint8)(2 | VRAM_OFFSET(1))	//static
+#define VRAM_D_0x06000000_ARM7 	(2 | VRAM_OFFSET(0))	//static
+#define VRAM_D_0x06020000_ARM7 	(2 | VRAM_OFFSET(1))	//static
 
 //todo VRAM setups 3-10, minus 7, minus 11, minus 3
